Minor fixing deleted old tabs
