+++ src/components/Tab/tab.js
@@ -2,11 +2,10 @@ import React from 'react';
 import classNames from 'classnames';
 import styles from './tab.scss';
 
-export const Tab2 = ({title, tabIndex, isActive, onClick}) => {
+export const Tab = ({title, tabIndex, isActive, onClick}) => {
   return (
     <a className={classNames('nav-item nav-link', styles.tab, {[`${styles.active}`]: isActive})} 
-      onClick={(event) => {
-        event.preventDefault();
+      onClick={() => {
         onClick(tabIndex);
       }}
     >{title}</a>
