+++ src/components/Tab/tab.scss
