--- src/components/Tab2/index.js
@@ -1,1 +1,0 @@
-export { Tab2 } from './tab.js';
