--- src/components/Tabs/tabs.scss
@@ -1,24 +1,0 @@
-@import '../../index.scss';
-
-.tab {
-  flex: 1;
-  @include tab;
-
-  &:not(:last-of-type) {
-    margin-right: -2%;
-  }
-
-  &:hover {
-    color: black;
-    background-color: #FFB800;
-
-    z-index: 3;
-  }
-}
-
-.active {
-  background-color:#FFB800;
-  font-weight: 600;
-  z-index: 2;
-}
-
