--- src/components/Tabs2/tabs.scss
@@ -1,6 +1,0 @@
-.active {
-  background-color:#FFB800;
-  font-weight: 600;
-  z-index: 2;
-}
-
