+++ src/components/index.js
@@ -1,6 +1,5 @@
 export { Tabs } from './tabs/index.js';
-export { Tabs2 } from './tabs2/index.js';
-export { Tab2 } from './tab2/index.js';
+export { Tab } from './tab/index.js';
 export { Button } from './Button/index.js';
 export { Card } from './card/index.js';
 export { Search } from './search/index.js';
